pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "braindb"
version = "0.1.0"
description = "Persistent memory for LLM agents — thoughts, facts, sources, and behavioral rules with fuzzy + semantic search, graph traversal, and an internal agent."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Dimitrios Koutsoumpos" },
]
keywords = [
    "llm",
    "agent",
    "memory",
    "knowledge-graph",
    "rag",
    "postgresql",
    "pgvector",
    "embeddings",
    "fastapi",
    "litellm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi==0.135.3",
    "uvicorn[standard]==0.44.0",
    "psycopg2-binary==2.9.11",
    "alembic==1.18.4",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "python-dotenv==1.2.2",
    "sentence-transformers==5.4.0",
    "numpy==2.4.4",
    "openai-agents[litellm]==0.13.6",
    "requests==2.33.1",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.3",
    "pytest-timeout==2.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["braindb"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 60
addopts = "-ra"
